fix driver close

+++ src/main/java/ui_automation/bolt/RecordHelper.java
@@ -156,6 +156,7 @@ public class RecordHelper {
     }
 
     public void recordMonthTripToDB(File monthTripCsv) {
+        if(!monthTripCsv.exists()) return;
         String content = IOUtils.readTextFromFile(monthTripCsv);
         monthTripCsv.deleteOnExit();
         List<String> rowArrayList = new ArrayList<>(Arrays.asList(content.split("\n")));
